% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{discrimination.scores}
\alias{discrimination.scores}
\title{Calculate the discrimination score for each cutoff}
\usage{
discrimination.scores(df, groups = c("label", "model.type"), divs = c(1,
  1.5, 2, 2.5, 3, 4, 6))
}
\arguments{
\item{df}{Data frame containing the score data from a set of models generated by Rosetta}

\item{disc.divs}{RMSD cutoffs to use for .}
}
\value{
A data.frame with the discrimination score for each target in the input data.frame.
}
\description{
Calculate the discrimination score for each cutoff
}
\examples{
d.scores <- discrimination.scores(score.data, divs = c(0.5, 1.0, 1.5, 2.0))
}

